Post install add symlinks for splash page.

+++ post_install.sh
@@ -21,9 +21,17 @@ if ! grep -q "wxt config_sync_directory" html/sites/default/settings.php; then
 fi
 
 cp custom/splash/.htaccess html/.htaccess
-ln -s custom/splash/splash.php html/splash.php
-ln -s custom/splash/sites/default/splash.js html/sites/default/splash.js
-ln -s custom/splash/sites/default/splash.css html/sites/default/splash.css
-ln -s custom/splash/sites/default/files/splashimages html/sites/default/files/splashimages
+if [ ! -L html/splash.php ]; then
+  ln -s custom/splash/splash.php html/splash.php
+fi
+if [ ! -L html/sites/default/splash.js ]; then
+  ln -s custom/splash/sites/default/splash.js html/sites/default/splash.js
+fi
+if [ ! -L html/sites/default/splash.css ]; then
+  ln -s custom/splash/sites/default/splash.css html/sites/default/splash.css
+fi
+if [ ! -L html/sites/default/files/splashimages ]; then
+  ln -s custom/splash/sites/default/files/splashimages html/sites/default/files/splashimages
+fi
 cp custom/splash/sites/default/*.png html/sites/default/.
